% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{carot_packages}
\alias{carot_packages}
\title{List all packages in the carot}
\usage{
carot_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include carot in the list?}
}
\description{
List all packages in the carot
}
\examples{
carot_packages()
}
