% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{compute_pca}
\alias{compute_pca}
\title{compute_pca}
\usage{
compute_pca(cohort_name, input_plink, output_directory, ref1kg_population)
}
\arguments{
\item{cohort_name}{A \code{character}.}

\item{input_plink}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{ref1kg_population}{A \code{character}.}
}
\description{
compute_pca
}
\keyword{internal}
