% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{estimate_ethnicity}
\alias{estimate_ethnicity}
\title{estimate_ethnicity}
\usage{
estimate_ethnicity(cohort_name, input_vcfs, input_type, output_directory,
  ref1kg_vcfs, ref1kg_population, ref1kg_maf = 0.05,
  splitted_by_chr = TRUE, quality_tag = "INFO",
  quality_threshold = 0.9, n_cores = 6, bin_path = list(vcftools =
  "/usr/bin/vcftools", bcftools = "/usr/bin/bcftools", bgzip =
  "/usr/bin/bgzip", tabix = "/usr/bin/tabix", plink1.9 =
  "/usr/bin/plink1.9"))
}
\arguments{
\item{cohort_name}{A \code{character}.}

\item{input_vcfs}{A \code{character}.}

\item{input_type}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{ref1kg_vcfs}{A \code{character}.}

\item{ref1kg_population}{A \code{character}.}

\item{ref1kg_maf}{A \code{numeric}.}

\item{splitted_by_chr}{A \code{logical}.}

\item{quality_tag}{A \code{character}.}

\item{quality_threshold}{A \code{numeric}.}

\item{n_cores}{An \code{integer}.}

\item{bin_path}{A \code{list(character)}.}
}
\value{
A \code{data.frame}.
}
\description{
estimate_ethnicity
}
