% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{format_array_chr}
\alias{format_array_chr}
\title{format_array_chr}
\usage{
format_array_chr(cohort_name, input_vcfs, output_directory, ref1kg_vcfs,
  ref1kg_maf, quality_tag, quality_threshold, n_cores, bin_path)
}
\arguments{
\item{cohort_name}{A \code{character}. A name to describe the studied population compared to 1,000 Genomes.}

\item{input_vcfs}{A \code{character}. A path to one or several VCFs file.}

\item{output_directory}{A \code{character}. The path where the data and figures is written.}

\item{ref1kg_vcfs}{A \code{character}. A path to the reference VCFs files (i.e., 1,000 Genomes sequencing data).}

\item{ref1kg_maf}{A \code{numeric}. MAF threshold for SNPs in 1,000 Genomes}

\item{quality_tag}{A \code{character}. Name of the imputation quality tag for \code{"array"}.}

\item{quality_threshold}{A \code{numeric}. The threshold to keep/discard SNPs based on their imputation quality.}

\item{n_cores}{An \code{integer}. The number of CPUs to use to estimate the ethnicity.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of
\code{vcftools}, \code{bcftools}, \code{bgzip}, \code{tabix} and \code{plink1.9}.}
}
\description{
format_array_chr
}
\keyword{internal}
