% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{format_sequencing}
\alias{format_sequencing}
\title{format_sequencing}
\usage{
format_sequencing(cohort_name, input_vcfs, output_directory, ref1kg_vcfs,
  ref1kg_maf, bin_path)
}
\arguments{
\item{cohort_name}{A \code{character}.}

\item{input_vcfs}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{ref1kg_vcfs}{A \code{character}.}

\item{ref1kg_maf}{A \code{numeric}.}

\item{bin_path}{A \code{list(character)}.}
}
\description{
format_sequencing
}
\keyword{internal}
