% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{format_vcf}
\alias{format_vcf}
\title{format_vcf}
\usage{
format_vcf(input_vcfs, ref1kg_vcfs, ref1kg_maf, ichr, quality_tag,
  quality_threshold, output_directory, bin_path)
}
\arguments{
\item{input_vcfs}{A \code{character}.}

\item{ref1kg_vcfs}{A \code{character}.}

\item{ref1kg_maf}{A \code{numeric}.}

\item{ichr}{A chracter or \code{numeric}.}

\item{quality_tag}{A \code{character}.}

\item{quality_threshold}{A \code{numeric}.}

\item{output_directory}{A \code{character}.}

\item{bin_path}{A \code{list(character)}.}
}
\description{
format_vcf
}
\keyword{internal}
