% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{ggheatmap}
\usage{
ggheatmap(data, dendrogram_columns = TRUE, dendrogram_rows = TRUE,
  line_size = 0.1, font_size = 11, font_size_x = 11,
  font_size_y = 11, distance_method = "euclidean",
  cluster_method = "ward.D2", scale_name = "value",
  scale_revert = FALSE)
}
\arguments{
\item{data}{description}

\item{dendrogram_columns}{description}

\item{dendrogram_rows}{description}

\item{line_size}{description}

\item{font_size}{description}

\item{font_size_x}{description}

\item{font_size_y}{description}

\item{distance_method}{description}

\item{cluster_method}{description}

\item{scale_name}{description}

\item{scale_revert}{description}
}
\value{
description
}
\description{
ggheatmap
}
