% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{merge_vcf}
\alias{merge_vcf}
\title{merge_vcf}
\usage{
merge_vcf(input_vcfs, bin_path)
}
\arguments{
\item{input_vcfs}{A \code{character}.}

\item{bin_path}{A \code{list(character)}.}
}
\description{
merge_vcf
}
\keyword{internal}
