% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_report.R
\name{pca_report}
\alias{pca_report}
\title{pca_report}
\usage{
pca_report(data, design, id_var = "Sample_ID", technical_vars,
  n_comp = 5, fig_n_comp = n_comp, outliers_component = NULL,
  outliers_threshold = 3, title_level = 2)
}
\arguments{
\item{data}{A \code{vector} or \code{data.frame}.}

\item{design}{A \code{data.frame}.}

\item{id_var}{A \code{character}.}

\item{technical_vars}{A \code{vector(character)}.}

\item{n_comp}{A \code{numeric}.}

\item{fig_n_comp}{A \code{numeric}.}

\item{outliers_component}{A \code{logical}.}

\item{outliers_threshold}{A \code{numeric}.}

\item{title_level}{A \code{numeric}.}
}
\value{
A \code{data.frame}.
}
\description{
pca_report
}
