% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mist.R
\name{print.mist}
\alias{print.mist}
\title{Print method for mist objec}
\usage{
\method{print}{mist}(x, ...)
}
\arguments{
\item{x}{[\link{mist}]}

\item{...}{Other arguments (Not used).}
}
\value{
list
}
\description{
Print method for mist objec
}
