% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_idats.R
\name{qc_idats}
\alias{qc_idats}
\title{qc_idats}
\usage{
qc_idats(csv_file, data_directory, output_file = "EPIC_QC.html",
  output_directory = ".", array = "EPIC",
  annotation = "ilm10b4.hg19", filter_snps = TRUE,
  filter_non_cpg = TRUE, filter_xy = TRUE, filter_multihit = TRUE,
  filter_beads = TRUE, population = NULL, bead_cutoff = 0.05,
  threshold_detection_pvalues = 0.01,
  threshold_call_rate_samples = 0.99, threshold_call_rate_cpgs = 1,
  threshold_gender = -2, colname_gender = "sexe",
  norm_background = "oob", norm_dye = "RELIC",
  norm_quantile = "quantile1", cell_tissue = NULL,
  pca_vars = c("Sample_Plate", "Sentrix_ID"),
  title = "EPIC Array Quality-Control", author_name = "CARoT",
  author_affiliation = "", author_email = "", show_code = FALSE,
  n_cores = 20, dpi = 300, gg_fontsize = 12, encoding = "UTF-8")
}
\arguments{
\item{csv_file}{A \code{character}.}

\item{data_directory}{A \code{character}.}

\item{output_file}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{array}{A \code{character}.}

\item{annotation}{A \code{character}.}

\item{filter_snps}{A \code{logical}.}

\item{filter_non_cpg}{A \code{logical}.}

\item{filter_xy}{A \code{logical}.}

\item{filter_multihit}{A \code{logical}.}

\item{filter_beads}{A \code{logical}.}

\item{population}{A \code{character}.}

\item{bead_cutoff}{A \code{numeric}.}

\item{threshold_detection_pvalues}{A \code{numeric}.}

\item{threshold_call_rate_samples}{A \code{numeric}.}

\item{threshold_call_rate_cpgs}{A \code{numeric}.}

\item{threshold_gender}{A \code{numeric}.}

\item{colname_gender}{A \code{character}.}

\item{norm_background}{A \code{character}.}

\item{norm_dye}{A \code{character}.}

\item{norm_quantile}{A \code{character}.}

\item{cell_tissue}{A \code{character}.}

\item{pca_vars}{A \code{vector(character)}.}

\item{title}{A \code{character}.}

\item{author_name}{A \code{character}.}

\item{author_affiliation}{A \code{character}.}

\item{author_email}{A \code{character}.}

\item{show_code}{A \code{logical}.}

\item{n_cores}{A \code{numeric}.}

\item{dpi}{A \code{numeric}.}

\item{gg_fontsize}{A \code{numeric}.}

\item{encoding}{A \code{character}.}
}
\description{
qc_idats
}
