% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_impute.R
\name{qc_impute}
\alias{qc_impute}
\title{Compute post-imputation quality-control report using a default rmarkdown template.}
\usage{
qc_impute(
  input_directory = NULL,
  output_directory = NULL,
  cohort_name = "CARoT",
  output_file = paste0(cohort_name, "_imputation_QC.html"),
  title = "Post-Imputation Quality-Control",
  author_name = "CARoT",
  author_affiliation = NULL,
  author_email = NULL,
  cache = FALSE,
  show_code = FALSE,
  n_cores = 1,
  dpi = 120,
  gg_fontsize = 12,
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{input_directory}{A \code{character}. The path to the VCF files.}

\item{output_directory}{A \code{character}. The path to the output directory.}

\item{cohort_name}{A \code{character}. The name of the studied cohort / population.}

\item{output_file}{A \code{character}. The name of the html file produced.}

\item{title}{A \code{character}. The report's title. Default is \code{"Post-Imputation Quality-Control"}.}

\item{author_name}{A \code{character}. The author's name to be printed in the report.
Default is \code{CARoT}.}

\item{author_affiliation}{A \code{character}. The affiliation to be printed in the report.
Default is \code{NULL}.}

\item{author_email}{A \code{character}. The email to be printed in the report.
Default is \code{NULL}.}

\item{cache}{A \code{logical}. Should the R code be cached?
Default is \code{FALSE}.}

\item{show_code}{A \code{logical}. Should the R code be printed in the report?
Default is \code{FALSE}.}

\item{n_cores}{A \code{numeric}. The number of CPUs to use to estimate the ethnicity.
Default is \code{1}.}

\item{dpi}{A \code{numeric}. The value for dpi when plotting the data.
Default is \code{120}.}

\item{gg_fontsize}{A \code{numeric}. Value for the font size. Default is \code{12}.}

\item{encoding}{A \code{character}. The encoding to be used for the html report.
Default is \code{"UTF-8"}.}

\item{...}{Parameters to pass to \code{rmarkdown::render()}.}
}
\description{
Compute post-imputation quality-control report using a default rmarkdown template.
}
