% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plink.R
\name{qc_plink}
\alias{qc_plink}
\title{qc_plink}
\usage{
qc_plink(input_directory = NULL, output_directory = NULL,
  cohort_name = NULL, output_file = paste0(cohort_name, "_QC.html"),
  array = NULL, callrate_samples = 0.95, callrate_snps = 0.95,
  heterozygosity_treshold = 4, maf_threshold = 0.01,
  hwe_pvalue = 1e-04, includes_relatives = FALSE,
  mendelian_samples = 0.05, mendelian_snp = 0.1, IBD_threshold = 0.2,
  population = NULL, pca_components = 10, pca_threshold = 3,
  check_bim_script = system.file("perl", "HRC-1000G-check-bim.pl",
  package = "CARoT"), ref1kg_panel = NULL, ref1kg_population = NULL,
  ref1kg_genotypes = NULL, ref1kg_legend = NULL, ref1kg_fasta = NULL,
  bin_path = list(bcftools = "/usr/bin/bcftools", bgzip =
  "/usr/bin/bgzip", plink = "/usr/bin/plink1.9", gcta = "/usr/bin/gcta64"),
  title = paste(array, "Array Quality-Control"), author_name = "CARoT",
  author_affiliation = NULL, author_email = NULL, cache = FALSE,
  show_code = FALSE, n_cores = 22, dpi = 120, gg_fontsize = 12,
  encoding = "UTF-8")
}
\arguments{
\item{input_directory}{A \code{character}.}

\item{output_directory}{A \code{character}. The path to the output directory.}

\item{cohort_name}{A \code{character}.}

\item{output_file}{A \code{character}. The name of the html file produced.}

\item{array}{A \code{character}. The array name.}

\item{callrate_samples}{A \code{numeric}.}

\item{callrate_snps}{A \code{numeric}.}

\item{heterozygosity_treshold}{A \code{numeric}.}

\item{maf_threshold}{A \code{numeric}.}

\item{hwe_pvalue}{A \code{numeric}.}

\item{includes_relatives}{A \code{logical}.}

\item{mendelian_samples}{A \code{numeric}.}

\item{mendelian_snp}{A \code{numeric}.}

\item{IBD_threshold}{A \code{numeric}.}

\item{population}{A \code{character}.}

\item{pca_components}{A \code{numeric}.}

\item{pca_threshold}{A \code{numeric}.}

\item{check_bim_script}{A \code{character}.}

\item{ref1kg_panel}{A \code{character}.}

\item{ref1kg_population}{A \code{character}.}

\item{ref1kg_genotypes}{A \code{character}.}

\item{ref1kg_legend}{A \code{character}.}

\item{ref1kg_fasta}{A \code{character}.}

\item{bin_path}{A \code{list(character)}.}

\item{title}{A \code{character}. The report's title. Default is \code{paste(array, "Array Quality-Control")}.}

\item{author_name}{A \code{character}. The author's name to be printed in the report.
Default is \code{CARoT}.}

\item{author_affiliation}{A \code{character}. The affiliation to be printed in the report.
Default is \code{NULL}.}

\item{author_email}{A \code{character}. The email to be printed in the report.
Default is \code{NULL}.}

\item{cache}{A \code{logical}. Should the R code be cached?
Default is \code{FALSE}.}

\item{show_code}{A \code{logical}. Should the R code be printed in the report?
Default is \code{FALSE}.}

\item{n_cores}{A \code{numeric}. The number of CPUs to use to estimate the ethnicity.
Default is \code{20}.}

\item{dpi}{A \code{numeric}. The value for dpi when plotting the data.
Default is \code{120}.}

\item{gg_fontsize}{A \code{numeric}. Value for the font size. Default is \code{12}.}

\item{encoding}{A \code{character}. The encoding to be used for the html report.
Default is \code{"UTF-8"}.}
}
\description{
qc_plink
}
