% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{Efficiently import idats files mostly using minfi functions.}
\usage{
read_idats(
  directory = getwd(),
  csv_file = "csv$",
  meth_value_type = "B",
  filter_beads = TRUE,
  bead_cutoff = 0.05,
  filter_non_cpg = TRUE,
  filter_snps = TRUE,
  population = NULL,
  filter_multihit = TRUE,
  filter_xy = TRUE,
  detection_pvalues = 0.01,
  filter_callrate = TRUE,
  callrate_samples = 0.99,
  callrate_probes = 1,
  norm_background = "oob",
  norm_dye = "RELIC",
  norm_quantile = "quantile1",
  array_name = c("EPIC", "450k"),
  annotation_version = c("ilm10b4.hg19", "ilmn12.hg19"),
  n_cores = 1,
  rgSet = NULL
)
}
\arguments{
\item{directory}{A \code{character}. Location of IDAT files, default is the current working directory.}

\item{csv_file}{A \code{character}. Path to the sample sheet (csv files) or
name of the sample sheet in \code{directory}.}

\item{meth_value_type}{A \code{character}. Indicates whether you prefer m-values (\code{"M"})
or beta-values (\code{"B"}). Default is \code{"B"}.}

\item{filter_beads}{A \code{logical}. Should the probes with a beadcount less than 3 be removed?
Default is \code{TRUE}.}

\item{bead_cutoff}{A \code{numeric}. The threshold for beadcount.
Default is \code{0.05}.}

\item{filter_non_cpg}{A \code{logical}. Should the non-cg probes be removed?}

\item{filter_snps}{A \code{logical}. Should the probes in which the probed CpG falls near a SNP
(according to (\href{https://www.doi.org/10.1093/nar/gkw967}{Zhou et al., 2016})) be removed?
Default is \code{TRUE}.}

\item{population}{A \code{character}. Name of the ethnicity population to be used.
Default is \code{NULL} for none.}

\item{filter_multihit}{A \code{logical}. Should the probes which align to multiple locations
(according to \href{https://www.doi.org/10.1186/gb-2013-14-9-r105}{Nordlund et al., 2013}) be removed?
Default is \code{TRUE}.}

\item{filter_xy}{A \code{logical}. Should the probes from X and Y chromosomes be removed?
Default is \code{TRUE}.}

\item{detection_pvalues}{A \code{numeric}. The threshold for the detection pvalues above which,
values are considered as missing. Default is \code{0.01}.}

\item{filter_callrate}{A \code{logical}. Should the data be filtered based on call rate metric?
Default is \code{TRUE}.}

\item{callrate_samples}{A \code{numeric}. The call rate threshold for samples, under which samples are excluded.
Default is \code{0.99}.}

\item{callrate_probes}{A \code{numeric}. The call rate threshold for probes, under which probes are excluded.
Default is \code{1}.}

\item{norm_background}{A \code{character}. Optional method to estimate background normal distribution parameters.
This must be one of \code{"oob"}, \code{"est"} or \code{"neg"}.
Default is \code{"oob"}.}

\item{norm_dye}{A \code{character}. Dye bias correction, "mean": correction based on averaged red/green ratio;
or \code{"RELIC"}: correction with RELIC method;
or \code{"none"}: no dye bias correction.
Default is \code{"RELIC"}.}

\item{norm_quantile}{A \code{character}. The quantile normalisation to be used.
This should be one of \code{"quantile1"}, \code{"quantile2"}, or \code{"quantile3"}.
Default is \code{"quantile1"}.}

\item{array_name}{A \code{character}. Choose microarray type, eiyther \code{"450K"} or \code{"EPIC"}.
Default is \code{"EPIC"}.}

\item{annotation_version}{A \code{character}. Version of the annotation package that should be used.
Default is \code{"ilm10b4.hg19"} for the \code{"EPIC"} array}

\item{n_cores}{An \code{integer}. The number of cores to use,
i.e., at most how many child processes will be run simultaneously.}

\item{rgSet}{A \code{RGChannelSet} object.}
}
\value{
A \code{list}.
}
\description{
Efficiently import idats files mostly using minfi functions.
}
