% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_metharray_exp}
\alias{read_metharray_exp}
\title{read_metharray_exp}
\usage{
read_metharray_exp(
  directory = NULL,
  sample_sheet = NULL,
  ignore.case = TRUE,
  recursive = TRUE,
  full.names = TRUE,
  n_cores = 1
)
}
\arguments{
\item{directory}{A \code{character}. Location of IDAT files, default is the current working directory.}

\item{ignore.case}{A \code{logical}. A logical value.
If \code{TRUE}, the directory path is prepended to the file names to give a relative file path.
If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{A \code{logical}. Should the listing recurse into directories?}

\item{full.names}{A \code{logical}. Should pattern-matching be case-insensitive?}

\item{n_cores}{An \code{integer}. The number of cores to use,
i.e., at most how many child processes will be run simultaneously.}
}
\description{
read_metharray_exp
}
