% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_sample_sheet}
\alias{read_sample_sheet}
\title{read_sample_sheet}
\usage{
read_sample_sheet(directory, csv_file = "csv$", ignore.case = TRUE,
  recursive = TRUE, full.names = TRUE)
}
\arguments{
\item{directory}{A \code{character}. Location of IDAT files,
default is the current working directory.}

\item{csv_file}{A \code{character}. Path to the sample sheet (csv files) or
name of the sample sheet in \code{directory}.}

\item{ignore.case}{A \code{logical}. A logical value.
If \code{TRUE}, the directory path is prepended to the file names to give a relative file path.
If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{A \code{logical}. Should the listing recurse into directories?}

\item{full.names}{A \code{logical}. Should pattern-matching be case-insensitive?}
}
\description{
read_sample_sheet
}
\keyword{internal}
