% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{format_array_chr}
\alias{format_array_chr}
\title{format_array_chr}
\usage{
format_array_chr(cohort_name, input_vcfs, output_directory, ref1kg_vcfs,
  ref1kg_maf, quality_tag, quality_threshold, n_cores, bin_path)
}
\arguments{
\item{cohort_name}{A \code{character}.}

\item{input_vcfs}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{ref1kg_vcfs}{A \code{character}.}

\item{ref1kg_maf}{A \code{numeric}.}

\item{quality_tag}{A \code{character}.}

\item{quality_threshold}{A \code{numeric}.}

\item{n_cores}{An \code{integer}.}

\item{bin_path}{A \code{list(character)}.}
}
\description{
format_array_chr
}
\keyword{internal}
