% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Compute heatmap with dendrogram on x-axis and y-axis using ggplot2.}
\usage{
ggheatmap(
  data,
  dendrogram_columns = TRUE,
  dendrogram_rows = TRUE,
  line_size = 0.1,
  font_size = 11,
  font_size_x = 11,
  font_size_y = 11,
  distance_method = "euclidean",
  cluster_method = "ward.D2",
  scale_name = "value",
  scale_revert = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}. The data on which to draw the heatmap/dendrogram.}

\item{dendrogram_columns}{A \code{logical}. Should the dendrogram for columns be plotted?}

\item{dendrogram_rows}{A \code{logical}. Should the dendrogram for row be plotted?}

\item{line_size}{A \code{numeric}. Value for the line size.}

\item{font_size}{A \code{numeric}. Value for the font size.}

\item{font_size_x}{A \code{numeric}. Value for the font size of the x-axis.}

\item{font_size_y}{A \code{numeric}. Value for the font size of the y-axis.}

\item{distance_method}{A \code{character}. The distance method to be used.
This must be one of \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.}

\item{cluster_method}{A \code{character}. The agglomeration method to be used.
This should be one of \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
\code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA), \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{scale_name}{A \code{character}. The name of the legend.}

\item{scale_revert}{A \code{logical}. Should the scale values be reversed?}
}
\value{
A \code{ggplot} object.
}
\description{
Compute heatmap with dendrogram on x-axis and y-axis using ggplot2.
}
