% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_report.R
\name{pca_report}
\alias{pca_report}
\title{pca_report}
\usage{
pca_report(data, design, id_var = "Sample_ID", technical_vars,
  n_comp = 5, fig_n_comp = n_comp, outliers_component = NULL,
  outliers_threshold = 3, title_level = 2)
}
\arguments{
\item{data}{A \code{vector} or \code{data.frame}. The data on which the PCA has to be performed.}

\item{design}{A \code{data.frame}. Additional variables to be used with factorial planes.}

\item{id_var}{A \code{character}. The identifier column used to merge the data.}

\item{technical_vars}{A \code{vector(character)}. Variables from design to be used with factorial planes.}

\item{n_comp}{A \code{numeric}. The number of principal components to be computed.}

\item{fig_n_comp}{A \code{numeric}. The number of principal components to be used for figures.}

\item{outliers_component}{A \code{logical}. The principal components to be used to outliers detection.}

\item{outliers_threshold}{A \code{numeric}. The treshold to define outliers.}

\item{title_level}{A \code{numeric}. The markdown title level, i.e., the number of \code{#} preceding the section.}
}
\value{
A \code{data.frame}.
}
\description{
pca_report
}
