% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_idats.R
\name{qc_idats}
\alias{qc_idats}
\title{qc_idats}
\usage{
qc_idats(csv_file, data_directory, array = "EPIC",
  annotation = "ilm10b4.hg19", output_file = paste(array, "QC.html"),
  output_directory = ".", filter_snps = TRUE, filter_non_cpg = TRUE,
  filter_xy = TRUE, filter_multihit = TRUE, filter_beads = TRUE,
  population = NULL, bead_cutoff = 0.05, detection_pvalues = 0.01,
  filter_callrate = TRUE, callrate_samples = 0.99,
  callrate_probes = 1, gender_threshold = -2, gender_colname = NULL,
  norm_background = "oob", norm_dye = "RELIC",
  norm_quantile = "quantile1", cell_tissue = NULL,
  pca_vars = c("Sample_Plate", "Sentrix_ID"), title = paste(array,
  "Array Quality-Control"), author_name = "CARoT",
  author_affiliation = NULL, author_email = NULL, cache = FALSE,
  show_code = FALSE, n_cores = 20, dpi = 300, gg_fontsize = 12,
  encoding = "UTF-8")
}
\arguments{
\item{csv_file}{A \code{character}.}

\item{data_directory}{A \code{character}.}

\item{array}{A \code{character}.}

\item{annotation}{A \code{character}.}

\item{output_file}{A \code{character}.}

\item{output_directory}{A \code{character}.}

\item{filter_snps}{A \code{logical}.}

\item{filter_non_cpg}{A \code{logical}.}

\item{filter_xy}{A \code{logical}.}

\item{filter_multihit}{A \code{logical}.}

\item{filter_beads}{A \code{logical}.}

\item{population}{A \code{character}.}

\item{bead_cutoff}{A \code{numeric}.}

\item{detection_pvalues}{A \code{numeric}.}

\item{filter_callrate}{A \code{logical}.}

\item{callrate_samples}{A \code{numeric}.}

\item{callrate_probes}{A \code{numeric}.}

\item{gender_threshold}{A \code{numeric}.}

\item{gender_colname}{A \code{character}.}

\item{norm_background}{A \code{character}.}

\item{norm_dye}{A \code{character}.}

\item{norm_quantile}{A \code{character}.}

\item{cell_tissue}{A \code{character}.}

\item{pca_vars}{A \code{vector(character)}.}

\item{title}{A \code{character}.}

\item{author_name}{A \code{character}.}

\item{author_affiliation}{A \code{character}.}

\item{author_email}{A \code{character}.}

\item{cache}{A \code{logical}.}

\item{show_code}{A \code{logical}.}

\item{n_cores}{A \code{numeric}.}

\item{dpi}{A \code{numeric}.}

\item{gg_fontsize}{A \code{numeric}.}

\item{encoding}{A \code{character}.}
}
\description{
qc_idats
}
