% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_idats.R
\name{qc_idats}
\alias{qc_idats}
\title{qc_idats}
\usage{
qc_idats(csv_file, data_directory, array = "EPIC",
  annotation = "ilm10b4.hg19", cohort_name = "CARoT",
  output_file = paste(cohort_name, array, "QC.html", sep = "_"),
  output_directory = ".", filter_snps = TRUE, filter_non_cpg = TRUE,
  filter_xy = TRUE, filter_multihit = TRUE, filter_beads = TRUE,
  population = NULL, bead_cutoff = 0.05, detection_pvalues = 0.01,
  filter_callrate = TRUE, callrate_samples = 0.99,
  callrate_probes = 1, gender_threshold = -2, gender_colname = NULL,
  norm_background = "oob", norm_dye = "RELIC",
  norm_quantile = "quantile1", cell_tissue = NULL,
  pca_vars = c("Sample_Plate", "Sentrix_ID"), title = paste(array,
  "Array Quality-Control"), author_name = "CARoT",
  author_affiliation = NULL, author_email = NULL, cache = FALSE,
  show_code = FALSE, n_cores = 1, dpi = 120, gg_fontsize = 12,
  encoding = "UTF-8", ...)
}
\arguments{
\item{csv_file}{A \code{character}. The path to a CSV file, \emph{i.e.}, a sample sheet describing the data.}

\item{data_directory}{A \code{character}. The path to the data directory.}

\item{array}{A \code{character}. The array name, \emph{i.e.}, \code{"EPIC"} or \code{"450k"}.}

\item{annotation}{A \code{character}. The name and version of the annotation package to be used.}

\item{cohort_name}{A \code{character}. The name of the studied cohort / population.}

\item{output_file}{A \code{character}. The name of the html file produced.}

\item{output_directory}{A \code{character}. The path to the output directory.}

\item{filter_snps}{A \code{logical}. Should the probes in which the probed CpG falls near a SNP
(according to (\href{https://www.doi.org/10.1093/nar/gkw967}{Zhou et al., 2016})) be removed?
Default is \code{TRUE}.}

\item{filter_non_cpg}{A \code{logical}. Should the non-cg probes be removed?}

\item{filter_xy}{A \code{logical}. Should the probes from X and Y chromosomes be removed?
Default is \code{TRUE}.}

\item{filter_multihit}{A \code{logical}. Should the probes which align to multiple locations
(according to \href{https://www.doi.org/10.1186/gb-2013-14-9-r105}{Nordlund et al., 2013}) be removed?
Default is \code{TRUE}.}

\item{filter_beads}{A \code{logical}. Should the probes with a beadcount less than 3 be removed?
Default is \code{TRUE}.}

\item{population}{A \code{character}. Name of the ethnicity population to be used.
Default is \code{NULL} for none.}

\item{bead_cutoff}{A \code{numeric}. The threshold for beadcount.
Default is \code{0.05}.}

\item{detection_pvalues}{A \code{numeric}. The threshold for the detection pvalues above which,
values are considered as missing. Default is \code{0.01}.}

\item{filter_callrate}{A \code{logical}. Should the data be filtered based on call rate metric?
Default is \code{TRUE}.}

\item{callrate_samples}{A \code{numeric}. The call rate threshold for samples, under which samples are excluded.
Default is \code{0.99}.}

\item{callrate_probes}{A \code{numeric}. The call rate threshold for probes, under which probes are excluded.
Default is \code{1}.}

\item{gender_threshold}{A \code{numeric}. The threshold value to discrimate gender based
on sexual chromosomes methylation.
Default is \code{-2}.}

\item{gender_colname}{A \code{character}. The name of the column containing the gender
in the file provided in \code{csv_file}.
Default is \code{NULL}.}

\item{norm_background}{A \code{character}. Optional method to estimate background normal distribution parameters.
This must be one of \code{"oob"}, \code{"est"} or \code{"neg"}.
Default is \code{"oob"}.}

\item{norm_dye}{A \code{character}. Dye bias correction, "mean": correction based on averaged red/green ratio;
or \code{"RELIC"}: correction with RELIC method;
or \code{"none"}: no dye bias correction.
Default is \code{"RELIC"}.}

\item{norm_quantile}{A \code{character}. The quantile normalisation to be used.
This should be one of \code{"quantile1"}, \code{"quantile2"}, or \code{"quantile3"}.
Default is \code{"quantile1"}.}

\item{cell_tissue}{A \code{character}. The cell tissue to be used for cell composition estimation,
using a reference panel (\emph{i.e.}, \code{"blood"} or \code{"cordblood"}) or a mathematical deconvolution.
Default is \code{NULL}.}

\item{pca_vars}{A \code{vector(character)}. Variables to be used with factorial planes.
Default is \code{c("Sample_Plate", "Sentrix_ID")}.}

\item{title}{A \code{character}. The report's title. Default is \code{paste(array, "Array Quality-Control")}.}

\item{author_name}{A \code{character}. The author's name to be printed in the report.
Default is \code{CARoT}.}

\item{author_affiliation}{A \code{character}. The affiliation to be printed in the report.
Default is \code{NULL}.}

\item{author_email}{A \code{character}. The email to be printed in the report.
Default is \code{NULL}.}

\item{cache}{A \code{logical}. Should the R code be cached?
Default is \code{FALSE}.}

\item{show_code}{A \code{logical}. Should the R code be printed in the report?
Default is \code{FALSE}.}

\item{n_cores}{A \code{numeric}. The number of CPUs to use to estimate the ethnicity.
Default is \code{1}.}

\item{dpi}{A \code{numeric}. The value for dpi when plotting the data.
Default is \code{120}.}

\item{gg_fontsize}{A \code{numeric}. Value for the font size. Default is \code{12}.}

\item{encoding}{A \code{character}. The encoding to be used for the html report.
Default is \code{"UTF-8"}.}

\item{...}{Parameters to pass to \code{rmarkdown::render()}.}
}
\description{
qc_idats
}
