% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plink.R
\name{qc_plink}
\alias{qc_plink}
\title{qc_plink}
\usage{
qc_plink(input_directory = NULL, output_directory = NULL,
  cohort_name = NULL, output_file = paste0(cohort_name, "_QC.html"),
  array = NULL, callrate_samples = 0.95, callrate_snps = 0.95,
  heterozygosity_treshold = 4, maf_threshold = 0.01,
  hwe_pvalue = 1e-04, includes_relatives = FALSE,
  mendelian_samples = 0.05, mendelian_snp = 0.1, IBD_threshold = 0.2,
  population = NULL, pca_components = 10, pca_threshold = 3,
  check_bim_script = system.file("perl", "HRC-1000G-check-bim.pl",
  package = "CARoT"), ref1kg_panel = NULL, ref1kg_population = NULL,
  ref1kg_genotypes = NULL, ref1kg_legend = NULL, ref1kg_fasta = NULL,
  bin_path = list(bcftools = "/usr/bin/bcftools", bgzip =
  "/usr/bin/bgzip", plink = "/usr/bin/plink1.9", gcta = "/usr/bin/gcta64"),
  title = paste(array, "Array Quality-Control"), author_name = "CARoT",
  author_affiliation = NULL, author_email = NULL, cache = FALSE,
  show_code = FALSE, n_cores = 22, dpi = 120, gg_fontsize = 12,
  encoding = "UTF-8", ...)
}
\arguments{
\item{input_directory}{A \code{character}. The path to the plink files.
The path should contains the file name without the extension, i.e., without \code{*.bed}, \code{*.bim} or \code{*.fam}.}

\item{output_directory}{A \code{character}. The path to the output directory.}

\item{cohort_name}{A \code{character}. The name of the studied cohort / population.}

\item{output_file}{A \code{character}. The name of the html file produced.}

\item{array}{A \code{character}. The array name, e.g., "Illumina Omni2.5".}

\item{callrate_samples}{A \code{numeric}. The call rate threshold for samples, under which samples are excluded.
Default is \code{0.95}.}

\item{callrate_snps}{A \code{numeric}. The call rate threshold for probes, under which probes are excluded.
Default is \code{0.95}.}

\item{heterozygosity_treshold}{A \code{numeric}. The heterozygosity threshold for samples
(number of standard deviation from the mean), under/above which samples are excluded.
Default is \code{4}.}

\item{maf_threshold}{A \code{numeric}. The minor allele frequency under which variants are considered "rare".
Default is \code{0.01}.}

\item{hwe_pvalue}{A \code{numeric}. The p-value threshold for Hardy-Weinberg equilibrium test.
Default is \code{0.0001}.}

\item{includes_relatives}{A \code{logical}. Does the data contain related samples?
Default is \code{FALSE}.}

\item{mendelian_samples}{A \code{numeric}. The Mendel error rate threshold above which samples are excluded.
Default is \code{0.05}.}

\item{mendelian_snp}{A \code{numeric}. The Mendel error rate threshold above which variants are excluded.
Default is \code{0.1}.}

\item{IBD_threshold}{A \code{numeric}. The threshold for IBD (identical by descent) above which
samples are characterised as relatives.
Default is \code{0.2}.}

\item{population}{A \code{character}. The ethnicity of the studied population if known, e.g., \code{"EUR"}.
Default is \code{NULL}.}

\item{pca_components}{A \code{numeric}. The number of principal components to be computed.
Default is \code{10}.}

\item{pca_threshold}{A \code{numeric}. The threshold to define outliers on the principal component analysis,
the as number of standard deviation from the cohort centroid.
Default is \code{3}.}

\item{check_bim_script}{A \code{character}. The PERL script to use to check PLINK files to allow later imputation.
Default is \code{system.file("perl", "HRC-1000G-check-bim.pl", package = "CARoT")}.}

\item{ref1kg_panel}{A \code{character}. The \code{*.panel} file from 1,000 Genome project.
Default is \code{NULL}.}

\item{ref1kg_population}{A \code{character}. The \code{*.tsv} file from 1,000 Genome project describing samples and ethnicity.
Default is \code{NULL}.}

\item{ref1kg_genotypes}{A \code{character}. The PLINK files from 1,000 Genome project.
Default is \code{NULL}.}

\item{ref1kg_legend}{A \code{character}. The \code{*.legend} file from 1,000 Genome project.
Default is \code{NULL}.}

\item{ref1kg_fasta}{A \code{character}. The \code{*.fasta} file from 1,000 Genome project.
Default is \code{NULL}.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of \code{bcftools}, \code{bgzip}, \code{gcta} and \code{plink1.9}.}

\item{title}{A \code{character}. The report's title. Default is \code{paste(array, "Array Quality-Control")}.}

\item{author_name}{A \code{character}. The author's name to be printed in the report.
Default is \code{CARoT}.}

\item{author_affiliation}{A \code{character}. The affiliation to be printed in the report.
Default is \code{NULL}.}

\item{author_email}{A \code{character}. The email to be printed in the report.
Default is \code{NULL}.}

\item{cache}{A \code{logical}. Should the R code be cached?
Default is \code{FALSE}.}

\item{show_code}{A \code{logical}. Should the R code be printed in the report?
Default is \code{FALSE}.}

\item{n_cores}{A \code{numeric}. The number of CPUs to use to estimate the ethnicity.
Default is \code{20}.}

\item{dpi}{A \code{numeric}. The value for dpi when plotting the data.
Default is \code{120}.}

\item{gg_fontsize}{A \code{numeric}. Value for the font size. Default is \code{12}.}

\item{encoding}{A \code{character}. The encoding to be used for the html report.
Default is \code{"UTF-8"}.}

\item{...}{Parameters to pass to \code{rmarkdown::render()}.}
}
\description{
qc_plink
}
