% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{read_idats}
\usage{
read_idats(directory = getwd(), csv_file = "csv$",
  meth_value_type = "B", filter_beads = TRUE, bead_cutoff = 0.05,
  filter_non_cpg = TRUE, filter_snps = TRUE, population = NULL,
  filter_multihit = TRUE, filter_xy = TRUE, array_name = c("EPIC",
  "450k"), annotation_version = c("ilm10b4.hg19", "ilmn12.hg19"),
  n_cores = 1)
}
\arguments{
\item{directory}{A \code{character}. Location of IDAT files,
default is the current working directory.}

\item{csv_file}{A \code{character}. Path to the sample sheet (csv files) or
name of the sample sheet in \code{directory}.}

\item{meth_value_type}{A \code{character}. Indicates whether you prefer m-values (\code{"M"})
or beta-values (\code{"B"}). Default is \code{"B"}.}

\item{filter_beads}{A \code{logical}. Should probes with a beadcount less than 3 will be removed
depending on the \code{bead_cutoff} value. Default is \code{TRUE}.}

\item{bead_cutoff}{A \code{numeric}. The beadCutoff represents the fraction of samples that must
have a beadcount less than 3 before the probe is removed.(default = 0.05)}

\item{filter_non_cpg}{A \code{logical}. Should non-cg probes be removed? Default is \code{TRUE}.}

\item{filter_snps}{A \code{logical}. Should probes in which the probed CpG falls near a
SNP be removed? Default is \code{TRUE}.}

\item{population}{A \code{vector(character)}.}

\item{filter_multihit}{A \code{logical}. Should probes in which the probe aligns to multiple
locations with bwa be removed? Default is \code{TRUE}.}

\item{filter_xy}{A \code{logical}. Should probes from X and Y chromosomes be removed? Default is \code{TRUE}.}

\item{array_name}{A \code{character}. Choose microarray type, eiyther \code{"450K"} or \code{"EPIC"}.
Default is \code{"EPIC"}.}

\item{annotation_version}{A \code{character}. Version of the annotation package that should be used.
Default is \code{"ilm10b4.hg19"} for the \code{"EPIC"} array}

\item{n_cores}{A \code{integer}. The number of cores to use,
i.e., at most how many child processes will be run simultaneously.}
}
\value{
A \code{list}.
}
\description{
read_idats
}
