% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_metharray}
\alias{read_metharray}
\title{read_metharray}
\usage{
read_metharray(basenames)
}
\arguments{
\item{basenames}{The \code{basenames} or \code{filenames} of the IDAT files.
\code{basenames} are the filename without the ending \verb{_Grn.idat} or \verb{_Red.idat}.
\code{filenames} are filenames including \verb{_Grn.idat} or \verb{_Red.idat}.}
}
\description{
read_metharray
}
\keyword{internal}
